A couple small changes to fix compiler warnings

+++ jawengine/types.h
@@ -37,10 +37,10 @@ namespace jaw {
 		int16_t x, y;
 		constexpr vec2i(int16_t x = 0, int16_t y = 0) { this->x = x; this->y = y; }
 
-		constexpr vec2i operator+(const vec2i rhs) const { return { x + rhs.x, y + rhs.y }; }
-		constexpr vec2i operator+(const int16_t rhs) const { return { x + rhs, y + rhs }; }
-		constexpr vec2i operator-(const vec2i rhs) const { return { x - rhs.x, y - rhs.y }; }
-		constexpr vec2i operator*(const int16_t rhs) const { return { x * rhs, y * rhs }; }
+		constexpr vec2i operator+(const vec2i rhs) const { return { (int16_t)(x + rhs.x), (int16_t)(y + rhs.y) }; }
+		constexpr vec2i operator+(const int16_t rhs) const { return { (int16_t)(x + rhs), (int16_t)(y + rhs) }; }
+		constexpr vec2i operator-(const vec2i rhs) const { return { (int16_t)(x - rhs.x), (int16_t)(y - rhs.y) }; }
+		constexpr vec2i operator*(const int16_t rhs) const { return { (int16_t)(x * rhs), (int16_t)(y * rhs) }; }
 		constexpr vec2i operator*(const float rhs) const { return { (int16_t)(x * rhs), (int16_t)(y * rhs) }; }
 		constexpr vec2i operator/(const float rhs) const { return { (int16_t)(x / rhs), (int16_t)(y / rhs) }; }
